+++ kimchi/src/circuits/polynomials/foreign_field_mul/gadget.rs
@@ -8,7 +8,8 @@ use crate::{
     alphas::Alphas,
     circuits::{
         argument::Argument,
-        expr::{Cache, E},
+        berkeley_columns::E,
+        expr::Cache,
         gate::{CircuitGate, GateType},
         lookup::{
             self,
